Function sgx_tunittest::rsgx_unit_test

pub fn rsgx_unit_test(
    ncases: &mut u64, 
    failurecases: &mut Vec<String>, 
    f: &'static Fn(), 
    name: &str
)

Perform one test case at a time.

This is the core function of sgx_tunittest. It runs one test case at a time and saves the result. On test passes, it increases the passed counter and on test fails, it records the failed test. Required test function must be Fn(), taking nothing as input and returns nothing.